Rename services/parseDna to services/parse_dna
Previously, parseDna was the only lowerCamelCased js fiename. Now
it matches the other filenames which are snake_case.
